# Conflicts:
#	projects/FEM/INT/src/CM/cl_FEM_CM_Compressible_Newtonian_Fluid.cpp
#	projects/FEM/INT/src/CM/cl_FEM_CM_Compressible_Newtonian_Fluid.hpp
#	projects/FEM/INT/src/CM/cl_FEM_Constitutive_Model.cpp
#	projects/FEM/INT/src/CM/cl_FEM_Constitutive_Model.hpp
#	projects/FEM/INT/src/CORE/cl_FEM_Set_User_Info.hpp
#	projects/FEM/INT/src/ELEM/cl_FEM_Cluster.hpp
#	projects/FEM/INT/src/IWG/cl_FEM_IWG.hpp
#	projects/FEM/INT/src/IWG/cl_FEM_IWG_Struc_Stress.hpp
#	projects/MRS/CNT/src/cl_Parameter.hpp
#	projects/MRS/CNT/src/cl_Parameter_List.hpp
#	projects/MRS/CNT/src/cl_Validator.cpp
#	projects/MRS/IOS/src/cl_Library_IO.cpp
#	projects/PRM/ENM/src/cl_FEM_Enums.hpp
#	projects/PRM/ENM/src/cl_NLA_Nonlinear_Solver_Enums.hpp
#	projects/PRM/ENM/src/cl_TSA_Time_Solver_Enums.hpp
#	projects/PRM/src/fn_PRM_FEM_Parameters.hpp
#	projects/PRM/src/fn_PRM_SOL_Parameters.hpp

+++ projects/SOL/DLA/src/cl_DLA_Linear_Solver_Amesos.cpp
@@ -228,7 +228,7 @@ Linear_Solver_Amesos::set_solver_internal_parameters()
     {
 #ifdef MORIS_USE_MUMPS
         // Increase memory allocation to 200% at a time (default = 20%)
-        params.sublist( "mumps" ).set( "ICNTL(14)", static_cast< int >( 200 ) );
+        params.sublist( "mumps" ).set( "ICNTL(14)", 200 );
 #else
         MORIS_ERROR( false,
                 "Linear_Solver_Amesos::set_solver_internal_parameters - MORIS installed without support for MUMPS" );
